completed fixed nav, updated test script to end if a command fails

+++ test-gen.sh
@@ -1,22 +1,24 @@
 #!/usr/bin/env bash
 
 echo compiling sass...
-npm run sass
-
-echo transpiling code...
-npm run babel
-
-if cd generated-example
+if npm run sass
 then
-  echo cleaning example folder...
-  rm -rdf ./*
-else
-echo createing example folder...
- mkdir generated-example
- cd generated-example
+  echo transpiling code...
+  if npm run babel
+  then
+    if cd generated-example
+    then
+      echo cleaning example folder...
+      rm -rdf ./*
+    else
+    echo createing example folder...
+     mkdir generated-example
+     cd generated-example
+    fi
+
+    echo generating wireframe...
+    yo html-wireframe
+
+    cd ..
+  fi
 fi
-
-echo generating wireframe...
-yo html-wireframe
-
-cd ..
